+++ proyect2/proyecto2/src/App.css
@@ -1,0 +1,71 @@
+#root {
+  max-width: 1280px;
+  margin: 0 auto;
+  padding: 2rem;
+  text-align: center;
+}
+
+.logo {
+  height: 6em;
+  padding: 1.5em;
+  will-change: filter;
+  transition: filter 300ms;
+}
+.logo:hover {
+  filter: drop-shadow(0 0 2em #646cffaa);
+}
+.logo.react:hover {
+  filter: drop-shadow(0 0 2em #61dafbaa);
+}
+
+@keyframes logo-spin {
+  from {
+    transform: rotate(0deg);
+  }
+  to {
+    transform: rotate(360deg);
+  }
+}
+
+@media (prefers-reduced-motion: no-preference) {
+  a:nth-of-type(2) .logo {
+    animation: logo-spin infinite 20s linear;
+  }
+}
+
+.card {
+  padding: 2em;
+}
+
+.read-the-docs {
+  color: #888;
+}
+
+.Logos{
+  display: flex;
+  justify-content: space-around;
+  align-items: center;
+  flex-wrap: wrap;
+
+}
+
+.nombrep{
+  font-size: 20px;
+  font-weight: bold;
+  align-content: center;
+  padding: 1rem;
+  background-color: lightblue;
+  border-radius: 1rem;
+  margin: 1rem;
+}
+.espa{
+  padding: 0.5rem;
+  margin: 0.5rem;
+}
+
+.label{
+  background-color: green;
+  color: white;
+  padding: 0.5rem;
+  border-radius: 5rem;
+}
