+++ proyect2/proyecto2/src/componentes/ejercicio1.txt
@@ -1,0 +1,16 @@
+1.- Crear Componente 
+
+2. Recibir una prop con el año actual
+
+3.- Usar funciones para sacar el año
+
+4.- Usar estados y eventos para tener dos botones.
+    -Pasar al proximo año 
+    -Ir al año anterior
+    -Mostrar en todo momento el año por pantalla
+
+5.- Validar que el año sea un número y que sea requerido(Prop)
+
+6.- Cambiar el año mediante un input de texto y que se cambie 
+dinamicamente.
+
