--- src/App.css
@@ -1,69 +1,0 @@
-#root {
-  max-width: 1280px;
-  margin: 0 auto;
-  padding: 2rem;
-  text-align: center;
-}
-
-.logo {
-  height: 6em;
-  padding: 1.5em;
-  will-change: filter;
-  transition: filter 300ms;
-}
-.logo:hover {
-  filter: drop-shadow(0 0 2em #646cffaa);
-}
-.logo.react:hover {
-  filter: drop-shadow(0 0 2em #61dafbaa);
-}
-
-@keyframes logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
-}
-
-@media (prefers-reduced-motion: no-preference) {
-  a:nth-of-type(2) .logo {
-    animation: logo-spin infinite 5s linear;
-  }
-}
-
-.card {
-  padding: 2em;
-}
-
-.read-the-docs {
-  color: #888;
-}
-
-.caja {
-    display: inline-flex;
-    width: 300px;
-    height: 300px;
-    background-color: white;
-    border-radius: 10px;
-    margin: 10px;
-    padding: 10px;
-    transition: all 300ms ease-in-out;
-    border: 2px solid black;
-    flex-direction: row;
-    flex-wrap: wrap;
-    justify-content: space-around;
-}
-
-.caja:hover {
-  box-shadow: -4px 10px 14px 6px rgba(0, 0, 0, 0.2);
-  transform: scale(1.01);
-}
-.caja__texto-caja
-{
-  color: black;
-  font-size: 20px;
-  font-weight: bold;
-  align-content: center;
-}
